Refactoring test to whitebox
